[s2p] handle case of panchro-only datasets

+++ app/s2p/scripts/helper_convert_outputs.sh
@@ -6,11 +6,16 @@ export PATH=/usr/bin:/usr/local/bin:/bin:$PATH
 BINBIN=`dirname \`which s2p.py\``/bin
 export PATH=$BINBIN:$PATH
 
-iion  s2p_results/roi_color_ref.tif s2p_results/roi_color_ref_preview.png
 qauto s2p_results/roi_ref.tif s2p_results/roi_ref_preview.png
 qauto s2p_results/dem.tif s2p_results/dem_preview.png
 bin2asc s2p_results/cloud.ply > cloud_ascii.ply
 
+# some datasets don't have color images
+if [ -f s2p_results/roi_color_ref.tif ]; then
+    qauto s2p_results/roi_color_ref.tif s2p_results/roi_color_ref_preview.png
+fi
+
+# some datasets have only pairs (no left right subfolders)
 if [ -d s2p_results/left ] ; then
     for f in s2p_results/{left,right}/tile_*
     do
@@ -33,5 +38,7 @@ fi
 
 # cp output files in the root output folder for archive
 cp s2p_results/dem_preview.png .
-cp s2p_results/roi_color_ref_preview.png .
 cp s2p_results/roi_ref_preview.png .
+if [ -f s2p_results/roi_color_ref_preview.png ]; then
+    cp s2p_results/roi_color_ref_preview.png .
+fi
